ui配下のpackage名をlowerCamelCaseから全て小文字に統一する
closes #110
