+++ app/src/main/java/jp/itIsNoMatter/routineTimerClone/ui/routinelist/RoutineListScreen.kt
@@ -1,4 +1,4 @@
-package jp.itIsNoMatter.routineTimerClone.ui.routineList
+package jp.itIsNoMatter.routineTimerClone.ui.routinelist
 
 import androidx.compose.foundation.combinedClickable
 import androidx.compose.foundation.layout.Column
