+++ app/src/main/java/jp/itIsNoMatter/routineTimerClone/ui/routinelist/RoutineListViewModel.kt
@@ -1,4 +1,4 @@
-package jp.itIsNoMatter.routineTimerClone.ui.routineList
+package jp.itIsNoMatter.routineTimerClone.ui.routinelist
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
